requirements.txt
django==3.1.1
djangorestframework==3.12.1
djoser==2.0.5